# [feature] Reduce memory allocations and increase rotation throughput