Cleared out a few very obvious things:

- Preallocation of slices with known size during DB writeout
- Reduce some assessments when populating `types.Key` (if IPv4 / IPv6 is known)
- Remove unnecessary `defer()` statement in `RunGroup` (to avoid putting function on stack)

Most importantly though I've switched the rotation logic to continuously operate on the _existing_ flow map instead of creating a duplicate each time (since we have to lock the map anyway there's no reason to do so). This not only reduces allocations but significantly improves performance of the whole operation (probably mostly _because_ there's less allocations). I've also validated that this has no negative impact on future flow insertions.

Some benchmarks:
```
           │ /tmp/plain.txt │          /tmp/prealloc.txt          │           /tmp/delete.txt           │         /tmp/noprealloc.txt         │
           │     sec/op     │    sec/op     vs base               │   sec/op     vs base                │   sec/op     vs base                │
Rotation-4     116.80m ± 1%   112.20m ± 0%  -3.94% (p=0.000 n=10)   93.57m ± 1%  -19.89% (p=0.000 n=10)   86.62m ± 1%  -25.84% (p=0.000 n=10)

           │ /tmp/plain.txt │          /tmp/prealloc.txt           │           /tmp/delete.txt            │         /tmp/noprealloc.txt          │
           │      B/op      │     B/op      vs base                │     B/op      vs base                │     B/op      vs base                │
Rotation-4     45.23Mi ± 0%   59.54Mi ± 0%  +31.63% (p=0.000 n=10)   51.83Mi ± 0%  +14.58% (p=0.000 n=10)   37.52Mi ± 0%  -17.05% (p=0.000 n=10)

           │ /tmp/plain.txt │         /tmp/prealloc.txt          │          /tmp/delete.txt           │        /tmp/noprealloc.txt         │
           │   allocs/op    │  allocs/op   vs base               │  allocs/op   vs base               │  allocs/op   vs base               │
Rotation-4      112.0k ± 0%   109.7k ± 0%  -2.12% (p=0.000 n=10)   105.7k ± 0%  -5.69% (p=0.000 n=10)   108.1k ± 0%  -3.56% (p=0.000 n=10)
```

where:

- `plain`: current `develop`
- `prealloc`: switch to preallocation of `hashmap` (has nothing to do with the aforementioned DB writeout)
- `delete`: operate on same `FlowMap` instead of duplication
- `noprealloc`: reversion of changes in `prealloc` since it's not beneficial when looking at the average case (initial benchmark was best-case scenario for that optimization)

As you can see, `transferAndAggregate()` now is about 25% faster while allocating significantly less (note also that the actual savings in allocations are much larger since those samples also include synthetic allocations from the benchmark itself).

Using a memory pool for the `hashmap` was not beneficial as it a) increased overall memory usage (because more buffers have to be kept and are less accessible to the garbage collector) and b) most of the allocations are triggered by bucket resizing, which are not affected by the memory pooling.